refactor: constants/colorにCOLORヲまとめた
